Add policies and authorize controllers, enable password resets functionality for users

+++ app/controllers/posts_controller.rb
@@ -29,6 +29,7 @@ class PostsController < ApplicationController
     def edit
         @post = Post.find_by(id: params[:id])
         @topic = @post.topic
+        authorize @post
     end
 
     def update
@@ -47,6 +48,7 @@ class PostsController < ApplicationController
     def destroy
         @post = Post.find_by(id: params[:id])
         @topic = @post.topic
+        authorize @post
 
         if @post.destroy
         	flash[:danger] = "Post deleted!"
